+++ app/bashCTRL/execShellCommand.js
