--- app/config/config.json
@@ -1,1 +1,0 @@
-{"phrasesPath":"./app/packages/newphrases.back.json","commandsPath":"./app/packages/newcommands.back.json","exactMatchThreshold":0.71,"closeMatchThreshold":0.65,"name":"Jarvis"}
