+++ app/configCTRL/config.json
@@ -1,0 +1,1 @@
+{"phrasesPath":"./app/packages/phrases.json","commandsPath":"./app/packages/commands.json","exactMatchThreshold":0.71,"closeMatchThreshold":0.65,"name":"Jarvis"}
