+++ app/index.html
@@ -6,7 +6,7 @@
 </head>
 <body>
   <div id="app"></div>
-  <script src="listener/speechRecognition.js"></script>
+  <script src="listenerCTRL/speechRecognition.js"></script>
   <script src="views/dist/bundle.js"></script>
 </body>
 </html>
