+++ app/listenerCTRL/listener.js
@@ -17,7 +17,6 @@ module.exports = function (cb, name, timeout) {
   listener.onresult = cb;
 
   listener.pause = function () {
-    console.log(this.name, 'has been paused');
     on = false;
     this.stop();
   };
@@ -39,8 +38,7 @@ module.exports = function (cb, name, timeout) {
   };
 
   listener.onstart = function (e) {
-    console.log(this.name, ' started!');
-    // console.log(this.switchListener.name, ' will start!');
+    console.log('listening');
     if (listener.hasTimeout) {
       this.selfDestruct();
       listener.hasTimeout = false;
@@ -51,7 +49,6 @@ module.exports = function (cb, name, timeout) {
   listener.switch = function () {
     on = false;
     listener.abort();
-    console.log(listener.name, ' has turned off');
     listener.switchListener.start();
   };
