+++ app/matchCTRL/matchingUtil.js
@@ -17,8 +17,14 @@ module.exports = function (actionPrefix, variable, commandsObj) {
   actionObj.userCommand = actionPrefix;
   actionObj.guessedCommand = null;
   actionObj.action = '';
-  var exactMatchThreshold = parseFloat(utils.get('exactMatchThreshold'));
-  var closeMatchThreshold = parseFloat(utils.get('closeMatchThreshold'));
+  if (process.env.NODE_ENV !== 'test') {
+    var exactMatchThreshold = parseFloat(utils.get('exactMatchThreshold'));
+    var closeMatchThreshold = parseFloat(utils.get('closeMatchThreshold'));
+  } else {
+    console.log('test');
+    var exactMatchThreshold = 0.8;
+    var closeMatchThreshold = 0.65;
+  }
 
   var phrases = commandsObj.phrases;
   var actions = commandsObj.rawCommands;
