+++ app/packages/commands.json
@@ -1,1 +1,1 @@
-{"dance":"say \"weeee\"","open the pod bay doors":"say \"I'm sorry dave, I'm afraid I can't do that\"","free willy":"say \"FREEEEEEDOOOOM\"","kyle cho pro tip":"say kyle cho pro tip: $(cat protips.txt | perl -MList::Util=shuffle -e 'print shuffle(<STDIN>);' | head -n 1)","directions to":"open https://www.google.com/maps/dir/<ARG del='+' chainkey='to'/>/<ARG del='+' />/","laugh":"say \"HAHAHAHa\""}
+{"dance":"say \"weeee\"","can you open the pod bay doors":"say \"I'm sorry dave, I'm afraid I can't do that\"","kyle cho pro tip":"say kyle cho pro tip: $(cat protips.txt | perl -MList::Util=shuffle -e 'print shuffle(<STDIN>);' | head -n 1)","directions to":"open https://www.google.com/maps/dir/<ARG del='+' chainkey='to'/>/<ARG del='+' />/"}
