+++ package.json
@@ -7,8 +7,8 @@
     "test": "./test"
   },
   "scripts": {
-    "test": "grunt test",
-    "start": "grunt",
+    "test": "NODE_ENV=test grunt test",
+    "start": "electron .",
     "view": "webpack --watch",
     "lint": "jscs ./"
   },
@@ -58,12 +58,14 @@
   "devDependencies": {
     "chai": "^3.4.1",
     "del": "^2.2.0",
+    "electron-prebuilt": "^0.35.4",
     "grunt": "^0.4.5",
     "grunt-docco": "^0.4.0",
     "grunt-mocha-test": "^0.12.7",
     "grunt-nodemon": "^0.4.1",
     "grunt-shell": "^1.1.2",
     "mocha": "^2.3.4",
+    "mock-localstorage": "^0.1.3",
     "node-sass": "^3.4.2",
     "node-stopwatch": "0.0.1",
     "sass-loader": "^3.1.2",
