--- rootPath.js
@@ -1,2 +1,0 @@
-'use strict';
-module.exports = process.env.NODE_ENV === 'production' ? __dirname : process.cwd();
